# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AndonLights",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/AndonLight": {
      "get": {
        "tags": [
          "AndonLight"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AndonLightDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AndonLightDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AndonLightDTO"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AndonLight"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "AndonLight"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLightDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLightDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLightDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/AndonLight/{id}": {
      "get": {
        "tags": [
          "AndonLight"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/AndonLight/{name}": {
      "post": {
        "tags": [
          "AndonLight"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonLightDTO"
                }
              }
            }
          }
        }
      }
    },
    "/daily": {
      "get": {
        "tags": [
          "State"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/monthly": {
      "get": {
        "tags": [
          "State"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/State": {
      "post": {
        "tags": [
          "State"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AndonStateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AndonStateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AndonStateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AndonStateDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonStateDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonStateDTO"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "State"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AndonStateDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AndonStateDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AndonStateDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/State/{id}": {
      "get": {
        "tags": [
          "State"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AndonStateDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonStateDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AndonStateDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AndonLightDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "stateCode": {
            "$ref": "#/components/schemas/LightStates"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "dateofCreation": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "AndonStateDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LightStates": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "StatsBaseDTO": {
        "type": "object",
        "properties": {
          "numberOfEntries": {
            "type": "integer",
            "format": "int32"
          },
          "minutesSpentInState": {
            "type": "number",
            "format": "double"
          },
          "dateOfStats": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "StatsResponseDTO": {
        "type": "object",
        "properties": {
          "greenStats": {
            "$ref": "#/components/schemas/StatsBaseDTO"
          },
          "yellowStats": {
            "$ref": "#/components/schemas/StatsBaseDTO"
          },
          "redStats": {
            "$ref": "#/components/schemas/StatsBaseDTO"
          }
        },
        "additionalProperties": false
      },
      "UpdateLightDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}